+++ api/mathsys/validate.py
@@ -15,23 +15,23 @@ router = APIRouter()
 
 # FUNCTION -> EXTENSIONS
 from website.extensions import (
-    binary as _binary,
     mathsys as _mathsys,
     post as _post,
+    response as _response
 )
 
 # FUNCTION -> DECLARATION
-@router.post("/api/mathsys/compile")
-async def output(request: Request) -> StreamingResponse:
+@router.post("/api/mathsys/validate")
+async def output(request: Request) -> JSONResponse:
     # DECLARATION -> ACTIVATE EXTENSIONS
-    binary = await _binary.namespace().init(request)
     mathsys = await _mathsys.namespace().init(request)
     post = await _post.namespace().init(request)
+    response = await _response.namespace().init(request)
     # DECLARATION -> ARGUMENT CHECKS
     if not post.checks(): raise HTTPException(**SUG.ERR[0])
     # DECLARATION -> ARGUMENT RELATIONSHIP
     if not post.exists("Mcode"): raise HTTPException(**SUG.ERR[1])
     # DECLARATION -> QUERY
     mathsys.load(post.data["Mcode"])
-    binary.load(mathsys.compile() if mathsys.validate else b"\x00")
-    return binary.get()
+    response.load(mathsys.validate)
+    return response.get()