+++ public/content/&/script.tsx
@@ -4,6 +4,10 @@
 
 // HEAD -> MODULES
 import * as $ from "$";
+import * as ß from "ß";
+
+// HEAD -> COMPONENTS
+import $Button from "ßButton";
 
 
 //
@@ -12,7 +16,7 @@ import * as $ from "$";
 
 // REMOVE -> FUNCTION
 export async function hide(): Promise<void> {
-    await $.inject($.SUG.ORG, <></>);
+    await ß.inject($.SUG.ORG, <></>);
 }
 
 
@@ -30,7 +34,7 @@ export async function show(): Promise<void> {
     $.setTitle("Abscissa");
     $.setDescription("Landing page.");
     // FUNCTION -> CONTENT
-    await $.inject($.SUG.ORG, 
+    await ß.inject($.SUG.ORG, 
         <>
             <h1>LeetCode but for math</h1>
             <h3>Learn math by solving real problems</h3>
@@ -38,6 +42,12 @@ export async function show(): Promise<void> {
                 type="button" 
                 value="Try it now"
                 onClick={async() => await $.redirect("/playground", true)}
+            />
+            <$Button
+                text="Try it now"
+                onClick={async() => await $.redirect("/playground", true)}
+                id="CTA"
+                tooltip=""
             />
         </>
     )